:recycle: [#516] Move checks to permission
